requirements.txt
twitchio==1.2.1
requests==2.24.0
requests-cache==0.5.2
requests_oauthlib==1.3.0
python-dotenv==0.14.0
websockets==8.1